+++ packages/plugin-infra/src/entry.ts
@@ -2,6 +2,7 @@ export * from './types.js'
 export * from './hooks/index.js'
 export { createPluginMessage, type DefaultPluginMessage, type PluginMessageEmitter } from './utils/message.js'
 export { createPluginRPC, createPluginRPCGenerator } from './utils/rpc.js'
+export * from './utils/getTabs.js'
 export * from './utils/getTabContent.js'
 export {
     getPluginDefine,
